07_LanGraph_FLow(Basic)/pyproject.toml
[project]
name = "lg-flow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Abdullah Iqbal", email = "hassangb544@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain-google-genai>=2.0.9",
    "langgraph>=0.2.70",
    "litellm>=1.60.8",
    "python-dotenv>=1.0.1",
]

[project.scripts]
lg-flow = "lg_flow.simple_flow:run_prompt_chaining_flow"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
